Day 7
Added some time functionality to the game loop. 1st steps to be able to control the frames per second the game will run at. So in this day in particular, we are just finding out, how many seconds/miliseconds it takes to run the game loop. Which will allow use to tell how many frames a sec in the game loop.